requirements.txt
transformers>=4.43.0
torch>=2.2.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pytest>=8.0.0
black>=24.1.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=4.25.0 